moved users/me into routers/auth
